% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{codify_df}
\alias{codify_df}
\title{Codify text with categories from a dataframe}
\usage{
codify_df(text, df_categories)
}
\arguments{
\item{text}{text to be tagged}

\item{categories}{dataframe with categories, patterns and colors}
}
\description{
given a text and a dataframe with categories, color the text with the categories.
}
