% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{collapse_categories}
\alias{collapse_categories}
\title{collapse categories}
\usage{
collapse_categories(DF)
}
\arguments{
\item{DF}{}
}
\description{
given a dataframe with categories, it collapses the categories into one
line in regex pattern. It assumes the first column is the category and the
second is the terms and pattern do be collapsed/united into one single line.
}
