% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{color_palette2}
\alias{color_palette2}
\title{A discrete color palette}
\format{
An object of class \code{character} of length 17.
}
\usage{
color_palette2
}
\description{
A discrete color palette
}
\keyword{datasets}
