% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{color_palette3}
\alias{color_palette3}
\title{A discrete color palette -3}
\format{
An object of class \code{character} of length 14.
}
\usage{
color_palette3
}
\description{
A discrete color palette -3
}
\keyword{datasets}
