% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_text.R
\name{color_text}
\alias{color_text}
\title{Giving a text, tags text with colors using a text pattern}
\usage{
color_text(text, pattern, color_txt = "yellow", mouseover_txt = "CodifyR")
}
\arguments{
\item{text}{text to be tagged}

\item{pattern}{text pattern to be tagged}

\item{color_txt}{color to be tagged}
}
\description{
Giving a text, tags text with colors using a text pattern
}
