% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{count_categories}
\alias{count_categories}
\title{count categories}
\usage{
count_categories(file)
}
\description{
from an html file, count the categories generated and returns some basic 
statistics, like frequency of each category in text, as well as the frequency
of terms from each category.
}
