% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_html.R
\name{build_html}
\alias{build_html}
\title{build html page from pure text}
\usage{
build_html(text, title_page = "codifyR")
}
\arguments{
\item{text}{text to be tagged}

\item{title_page}{title of the page}
}
\description{
build html page from pure text. It must be supplied a text, generally from color_text() function.
It is possible to supply also the title of the page that will be shown in web browser.
}
