% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{color_palette}
\alias{color_palette}
\title{A discrete color palette - 1}
\format{
An object of class \code{character} of length 15.
}
\usage{
color_palette
}
\description{
A discrete color palette - 1
}
\keyword{datasets}
