% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\docType{data}
\name{color_palette4}
\alias{color_palette4}
\title{A discrete color palette}
\format{
An object of class \code{character} of length 10.
}
\usage{
color_palette4
}
\description{
A discrete color palette
}
\keyword{datasets}
