% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{colorize}
\alias{colorize}
\title{Create a discrete color column}
\usage{
colorize(df_categories, colors = color_palette)
}
\arguments{
\item{DF}{a dataframe with categories}

\item{colors.}{A vector of colors. It is possible to use the pre built
color_palette, color_palette2 or color_palette3 or to create your own, using
.html color names or hex colors.}
}
\description{
For using with the dataframe with categories and pattern, this function
creates a column with colors.
}
